setting up partition
commit message is vaguely misleading, I drew the partition so that I know I'm doing it correctly, but I still need to actually create the partition. Everything else seems to be working, and when I draw the effective radius it's inside the nine boxes that I expected it to be, so I'm excited to see what happens with the partition. Hopefully it actually makes it run smoothly and I can increase the load of particles, but it would also be fun to experiment with using the GPU and/or different threads to increase particle load